+++ switch-vision-snmp2mqtt/src/vendors/cisco/discovery.ts
@@ -1,5 +1,0 @@
-/**
- * Switch Vision discovery helpers.
- */
-
-export {}