+++ switch-vision-snmp2mqtt/src/vendors/cisco/sensor.ts
@@ -1,5 +1,0 @@
-/**
- * Switch Vision sensor helpers.
- */
-
-export {}