+++ switch-vision-snmp2mqtt/src/vendors/cisco/stack.ts
@@ -1,5 +1,0 @@
-/**
- * Switch Vision stack helpers.
- */
-
-export {}